+++ src/main/java/taskaya/backend/entity/freelancer/FreelancerBusiness.java
@@ -28,7 +28,7 @@ public class FreelancerBusiness {
     private Integer completedJobs;
 
     @Column(nullable = false)
-    private Integer rate; // Should be validated to [1-5] in logic
+    private float rate; // Should be validated to [1-5] in logic
 
     public enum State {
         BUSY, AVAILABLE
